// types/appliance.ts
export interface Appliance {
  id: string;
  appliance: string;
  brandModel: string;
  serialNumber: string;
  modelNumber: string;
  purchaseDate: string;
  warrantyEnd: string;
  filterPartNumber: string;
  manualLink: string;
  lastService: string;
  purchasePrice: string;
  purchaseLink: string;
  repair1Cost: string;
  repair1Date: string;
}